test: move channel actor coverage into fiber test module
